`ask` and `choose` use their own TTY
This enables them to isolate the context of interactivity, yet keep the parent TTY history and context from being cleared afterwards.

The TTY cursor work was not successful, as if the current row is nearing the terminal's total viewport lines, then it falls apart. As such the alternative TTY was the option.

Unfortunatey, the alternative TTY option has no way of keeping the stderr, despite efforts. As such, stderr will sleep for 10 seconds instead before the exit clears. This allows the rare error condition to be visible.

The other change this does is fix/improve `set (+|-)e` handling, as if a child did `set -e` then the return would exit the program. As such, this restores to what it was prior, which allows returns to not exit.

The next commit will trim the testing code and commentary.

Signed-off-by: Benjamin Lupton <b@lupton.cc>