setup-ruby: better homebrew ruby support
On homebrew systems, there could by multiple ruby versions installed, as dependencies of other packages.
If this is the case, this commit forces the user to select a preference.

It also removes opt/llvm gem handling, not sure why that was there, perhaps a bug.
And simplifies code around when `RUBY_VERSION` is not specified yet, allowing graceful operation.

Compiler flags will only now be provided for the selecte `RUBY_VERSION` install.

Also achieves formatting changes for `setup-ruby`, so progress on #74