setup-util: fix always using fallbacks
/ref 35a495b09ea686e09a3b3aa0aa458677d6485677

This is a regression from the referenced commit, which fixed using fallbacks. However now that fallbacks are used, they were always used.

This fix is done by returning instead of breaking (as such had_break is had_finish) so one does not get confused. As well as checking for had_success = no.