docs: dorothy-not-loading: merge with balupton's settings
I've also removed the `-i` recommendation, as it is not needed (terminals will be interactive by default). Protip, with shortflags, e.g. `-i` and `-l`, they can be combined into `-il`.

Also used `nu` instead of `NuShell`, as `nu` is the proper name outside initial exposure, and corrected the JSON formatting.

Turns out this advice is useful for VSCode in general, so removed the WSL reference.