choose, ask: fixes and improvements /ref #229
`ask` and `choose` now use our own `echo-wrap` and `echo-count-lines` instead of `gfold`, `gfmt`, `gwc`.
The reasoning for this was that they did not handle ansi escape codes, including colours, nor indentation of broken lines correctly.
To achieve this, I have introduce `echo-gnu-command` to fetch a GNU command (or fallback), with optional install.
Calls to `echo-gnu-command` may need `--no-fallback` if it turns out they actually require the GNU variant.
Considering this, we may be able to remove the symlink alias inside `setup-util-gawk` and `setup-util-gsed`.

This also fixes an issue with `choose` where the right arrow (page down) was not working when on the last item of a page.

Finally, this improves the formatting of subsequent lines in choose, they now have a continious quote line.

I have also figreud out when the CI is showing truncated output for choose, it occurs with `paging_supported='no'` this will be resolved another day.