choose, echo-wrap, read-key: support no-TTY
This adds support for no-TTY, such as a CI environment, or via `ssh -T ...` to `choose`.

However, this rewrites `read-key` to support such, and there needs further work as it still sends multiple characters in a single line under certain circumstances.

However, `read-key` now does support no-TTY situations, such as pipes, or bash v3 character mashing.

The complexity with this, is that there is no way native to just read one key at a time, and escape sequences can take multiple characters, as such the approach taken is to trim known escape sequences and then split the remaining. This is fine, however it means read-key now only supports known characters, and as such is more adequate to be considered a read-keyboard-key or so, as special escape sequences will require a lot of complexity to support, as they will need say our regex handling in `echo-wrap.awk` to be trimmed.

This all means that `read-key` should be changed instead of reading one key at a time, now in an aggregate, it should just read everything and then shift the keys off.