init: properly enable `NVIM` support (#279)
Under neovim, `$0` is the shell that invoked neovim, so if zsh is used to invoke neovim then it would be `zsh`. This is because of `/bin/zsh -c 'echo $0'` is `/bin/zsh` and `zsh -c 'echo $0'` is `zsh`. As such, we need to do the NVIM check earlier, as otherwise it will fall into the other checks, as `BASH_VERSION` will exit is bash invoked and `ZSH_VERSION` will exist if zsh invoked.

Co-authored-by: Benjamin Lupton <b@lupton.cc>