bash.bash: fix `setup-environment-commands --debug`
It was due to `--debug` in `__dump --debug ...` being treated as a reference, rather than being forwarded to the target. As such, I've also added `--stderr` and forwarding of all `__dump` unrecognised options to `__print_style`. This paves way for inlining `__print_style` and using it instead of `__debug` perhaps.