setup-util-nerd-fonts: fixed it
It wasn't doing anything, as before, if one font was installed, none the rest would be. This one calls `setup-util` individually for each font, which is not performant, but it is what it is.

Ideally, there will be a `setup-util --only-print-action "$@"` so we can get the right terminology for teh `choose` prompt, whether it is install, update, or uninstall.