ansi.bash: inline `__read_key`, confirm: fix clearing
This commit fixes `confirm` not clearing after `enter` correctly on slow machines. This was due to the `read` call in `read-key` then calling `add`, which then delayed the subsequent `read` which could cause an echo as no `read` would be active. This now captures the readings, then processes it, to ensure that the entire read is captured, even on slow machines.

Furthermore, this also improves performance by inlining `read-key` into `ansi.bash` as `__read_key`.

The remaining significant delay is now with `echo-clear-lines` calling `echo-wrap`, however this is worked around somewhat by the new `ansi.bash:__should_wrap` helper.

Note, that this does not prevent multiple enter presses from stuffing uup the clearing. This will be worked around by using `stty -echo` and `stty +echo` in a future commit.

Finally, this inlines `is-number` to `bash.bash:__is_number`.