implement custom and proper `__print_help` parser
This handles edge cases with precision, erroring on ambiguous cases with error suggestion on appropriate syntax. It changes the `: ` workaround for actions to `| `, however that workaround may be unnecessary now. Will continue to experiment.

Eventually, this custom parser technique could be applied to inlining `echo-wrap` into `ansi.bash` eliminating the GNU dependencies.